pyproject.toml
[build-system]
requires = ["poetry-core>=1.2.0", "wheel",]
build-backend = "poetry.core.masonry.api"


[tool.coverage.report]
exclude_lines = [
    "pragma: not covered",
    "@overload",
]


[tool.coverage.run]
branch = true
relative_files = true
omit = [
    ".tox/*",
    "tests/*",
]


[[tool.mypy.overrides]]
module = "uvicorn.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "fastapi_mail.*"
ignore_missing_imports = true


[tool.poetry]
name = "portfolio_rest_api"
version = "0.1.0"
description = "My personal portfolio back-end."

packages = [
    { include = "portfolio_rest_api", from = "src" },
]

authors = ["Lari Liuhamo <lari.liuhamo+pypi@gmail.com>",]
maintainers = ["Lari Liuhamo <lari.liuhamo+pypi@gmail.com>",]

include = []
license = "MIT"
readme = "README.md"

homepage = "https://pypi.org/project/portfolio_rest_api/"
repository = "https://github.com/Diapolo10/portfolio_rest_api"
documentation = "https://github.com/Diapolo10/portfolio_rest_api/tree/main/docs"

keywords = [
    "python3",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Typing :: Typed",
]


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
fastapi-mail = "^1.4.1"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"


[tool.poetry.group.linters]
optional = true


[tool.poetry.group.linters.dependencies]
ruff = "^0.7.1"


[tool.poetry.group.tests]
optional = true


[tool.poetry.group.tests.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
tox = "^4.23.2"
tox-gh-actions = "^3.2.0"


[tool.poetry.urls]
"Source code" = "https://github.com/Diapolo10/portfolio_rest_api"
"Tracker" = "https://github.com/Diapolo10/portfolio_rest_api/issues"
"Changelog" = "https://github.com/Diapolo10/portfolio_rest_api/blob/main/CHANGELOG.md"


[tool.pytest.ini_options]
minversion = "6.0"
addopts = """
--doctest-modules \
--cov=./ \
--cov-append \
--cov-report html:tests/reports/coverage-html \
--cov-report xml:tests/reports/coverage.xml \
--ignore=docs/
"""
testpaths = [
    "tests",
]


[tool.ruff]
lint.select = [
    "A",     # Builtins
    "ANN",   # Annotations
    "ARG",   # Unused arguments
    "ASYNC", # Asynchronous code
    "B",     # Bugbear
    "BLE",   # Blind except
    "C4",    # Comprehensions
    "C90",   # mccabe
    "COM",   # Commas
    "D1",    # Undocumented public elements
    "D2",    # Docstring conventions
    "D3",    # Triple double quotes
    "D4",    # Docstring text format
    "DTZ",   # Datetimes
    "E",     # Errors
    "EM",    # Error messages
    "ERA",   # Commented-out code
    "EXE",   # Executable
    "F",     # Pyflakes
    "FA",    # __future__ annotations
    # "FAST",  # FastAPI
    "FBT",   # "Boolean trap"
    "FIX",   # "FIXME"-comments
    "FLY",   # F-strings
    "FURB",  # Refurb
    "G",     # Logging format
    "I",     # Isort
    "ICN",   # Import conventions
    "INP",   # Disallow PEP-420 (Implicit namespace packages)
    "INT",   # gettext
    "ISC",   # Implicit str concat
    "LOG",   # Logging
    "N",     # PEP-8 Naming
    "PERF",  # Unnecessary performance costs
    "PGH",   # Pygrep hooks
    "PIE",   # Unnecessary code
    "PL",    # Pylint
    "PT",    # Pytest
    "PTH",   # Use Pathlib
    "PYI",   # Stub files
    "Q",     # Quotes
    "RET",   # Return
    "RUF",   # Ruff
    "RSE",   # Raise
    "S",     # Bandit
    "SIM",   # Code simplification
    "SLF",   # Private member access
    "SLOT",  # __slots__
    "T10",   # Debugger
    "T20",   # Print
    "TCH",   # Type checking
    "TD",    # "TODO"-comments
    "TID",   # Tidy imports
    "TRY",   # Exception handling
    "UP",    # Pyupgrade
    "W",     # Warnings
    "YTT",   # sys.version
]
lint.ignore = [
    "ANN101",  # Type annotation for `self`
    "D203",    # One blank line before class docstring
    "D212",    # Multi-line summary first line
    "PLR0913", # Too many arguments
    "Q000",    # Single quotes found but double quotes preferred
]
line-length = 120
preview = true
show-fixes = true
src = ["src",]
target-version = "py312"


[tool.ruff.lint.flake8-copyright]
author = "Lari Liuhamo"


[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
multiline-quotes = "double"


[tool.ruff.lint.mccabe]
max-complexity = 10


[tool.ruff.lint.per-file-ignores]
# https://beta.ruff.rs/docs/rules/
"__init__.py" = ["F401", "F403", "F405",]
"tests/*" = ["ANN", "ARG", "INP001", "S101",]
"logger.py" = ["N815",]


[tool.ruff.lint.pylint]
max-args = 15
max-branches = 20
max-returns = 10
max-statements = 80


[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"


[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py312
skip_missing_interpreters = true

[gh-actions]
python =
    3.12: py312

[testenv]
passenv = GITHUB_*
allowlist_externals = poetry

commands =
    poetry run pytest
    poetry run coverage report
"""
